+++ Form/Type/CKEditorType.php
@@ -17,7 +17,8 @@ use Ivory\CKEditorBundle\Model\ConfigManagerInterface,
     Symfony\Component\Form\FormBuilderInterface,
     Symfony\Component\Form\FormView,
     Symfony\Component\Form\FormInterface,
-    Symfony\Component\OptionsResolver\OptionsResolverInterface;
+    Symfony\Component\OptionsResolver\OptionsResolverInterface,
+    Symfony\Component\Templating\Helper\CoreAssetsHelper;
 
 /**
  * CKEditor type.
@@ -29,25 +30,47 @@ class CKEditorType extends AbstractType
     /** @var boolean */
     protected $enable;
 
+    /** @var string */
+    protected $basePath;
+
+    /** @var string */
+    protected $jsPath;
+
     /** @var \Ivory\CKEditorBundle\Model\ConfigManagerInterface */
     protected $configManager;
 
     /** @var \Ivory\CKEditorBundle\Model\PluginManagerInterface */
     protected $pluginManager;
 
+    /** @var \Symfony\Component\Templating\Helper\CoreAssetsHelper */
+    protected $assetsHelper;
+
     /**
      * Creates a CKEditor type.
      *
-     * @param boolean                                            $enable        TRUE if you want to use ckeditor widget,
-     *                                                                          FALSE if you want to use textarea widget.
-     * @param \Ivory\CKEditorBundle\Model\ConfigManagerInterface $configManager The CKEditor config manager.
-     * @param \Ivory\CKEditorBundle\Model\PluginManagerInterface $pluginManager The CKEditor plugin manager.
+     * @param boolean                                               $enable        TRUE if you want to use ckeditor widget,
+     *                                                                             FALSE if you want to use textarea widget.
+     * @param string                                                $basePath      The CKEditor base path.
+     * @param string                                                $jsPath        The CKEditor JS path.
+     * @param \Ivory\CKEditorBundle\Model\ConfigManagerInterface    $configManager The CKEditor config manager.
+     * @param \Ivory\CKEditorBundle\Model\PluginManagerInterface    $pluginManager The CKEditor plugin manager.
+     * @param \Symfony\Component\Templating\Helper\CoreAssetsHelper $assetsHelper  The assets helper.
      */
-    public function __construct($enable, ConfigManagerInterface $configManager, PluginManagerInterface $pluginManager)
+    public function __construct(
+        $enable,
+        $basePath,
+        $jsPath,
+        ConfigManagerInterface $configManager,
+        PluginManagerInterface $pluginManager,
+        CoreAssetsHelper $assetsHelper
+    )
     {
         $this->isEnable($enable);
-        $this->configManager = $configManager;
-        $this->pluginManager = $pluginManager;
+        $this->setBasePath($basePath);
+        $this->setJsPath($jsPath);
+        $this->setConfigManager($configManager);
+        $this->setPluginManager($pluginManager);
+        $this->setAssetsHelper($assetsHelper);
     }
 
     /**
@@ -64,6 +87,46 @@ class CKEditorType extends AbstractType
         }
 
         return $this->enable;
+    }
+
+    /**
+     * Gets the CKEditor base path.
+     *
+     * @return string The CKEditor base path.
+     */
+    public function getBasePath()
+    {
+        return $this->basePath;
+    }
+
+    /**
+     * Sets the CKEditor base path.
+     *
+     * @param string $basePath The CKEditor base path.
+     */
+    public function setBasePath($basePath)
+    {
+        $this->basePath = $basePath;
+    }
+
+    /**
+     * Gets the CKEditor JS path.
+     *
+     * @return string The CKEditor JS path.
+     */
+    public function getJsPath()
+    {
+        return $this->jsPath;
+    }
+
+    /**
+     * Sets the CKEditor JS path.
+     *
+     * @param string $jsPath The CKEditor JS path.
+     */
+    public function setJsPath($jsPath)
+    {
+        $this->jsPath = $jsPath;
     }
 
     /**
@@ -107,6 +170,26 @@ class CKEditorType extends AbstractType
     }
 
     /**
+     * Gets the assets helper.
+     *
+     * @return \Symfony\Component\Templating\Helper\CoreAssetsHelper The assets helper.
+     */
+    public function getAssetsHelper()
+    {
+        return $this->assetsHelper;
+    }
+
+    /**
+     * Sets the assets helper.
+     *
+     * @param \Symfony\Component\Templating\Helper\CoreAssetsHelper $assetsHelper The assets helper.
+     */
+    public function setAssetsHelper(CoreAssetsHelper $assetsHelper)
+    {
+        $this->assetsHelper = $assetsHelper;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
@@ -114,8 +197,10 @@ class CKEditorType extends AbstractType
         $builder->setAttribute('enable', $options['enable']);
 
         if ($builder->getAttribute('enable')) {
-            $config = $options['config'];
+            $builder->setAttribute('base_path', $this->assetsHelper->getUrl($options['base_path']));
+            $builder->setAttribute('js_path', $this->assetsHelper->getUrl($options['js_path']));
 
+            $config = $options['config'];
             if ($options['config_name'] === null) {
                 $name = uniqid('ivory', true);
 
@@ -140,6 +225,9 @@ class CKEditorType extends AbstractType
         $view->vars['enable'] = $form->getConfig()->getAttribute('enable');
 
         if ($form->getConfig()->getAttribute('enable')) {
+            $view->vars['base_path'] = $form->getConfig()->getAttribute('base_path');
+            $view->vars['js_path'] = $form->getConfig()->getAttribute('js_path');
+
             $view->vars['config'] = $form->getConfig()->getAttribute('config');
             $view->vars['plugins'] = $form->getConfig()->getAttribute('plugins');
         }
@@ -153,6 +241,8 @@ class CKEditorType extends AbstractType
         $resolver->setDefaults(array(
             'required'    => false,
             'enable'      => $this->enable,
+            'base_path'   => $this->basePath,
+            'js_path'     => $this->jsPath,
             'config_name' => null,
             'config'      => array(),
             'plugins'     => array(),
@@ -161,6 +251,8 @@ class CKEditorType extends AbstractType
         $resolver->addAllowedTypes(array(
             'enable'      => 'bool',
             'config_name' => array('string', 'null'),
+            'base_path'   => array('string'),
+            'js_path'     => array('string'),
             'config'      => 'array',
             'plugins'     => 'array',
         ));